@@ -11,15 +11,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^21.0.3",
-    "@angular/common": "^21.0.3",
-    "@angular/compiler": "^21.0.3",
-    "@angular/core": "^21.0.3",
-    "@angular/forms": "^21.0.3",
-    "@angular/platform-browser": "^21.0.3",
-    "@angular/platform-browser-dynamic": "^21.0.3",
-    "@angular/router": "^21.0.3",
-    "@babel/runtime": "^7.26.7",
+    "@angular/animations": "^21.1.3",
+    "@angular/common": "^21.1.3",
+    "@angular/compiler": "^21.1.3",
+    "@angular/core": "^21.1.3",
+    "@angular/forms": "^21.1.3",
+    "@angular/platform-browser": "^21.1.3",
+    "@angular/platform-browser-dynamic": "^21.1.3",
+    "@angular/router": "^21.1.3",
+    "@babel/runtime": "^7.26.10",
     "@emailjs/browser": "^3.11.0",
     "@ngx-translate/core": "^13.0.0",
     "@ngx-translate/http-loader": "^6.0.0",
@@ -32,11 +32,11 @@
     "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular/build": "^21.0.2",
-    "@angular/cli": "^21.0.2",
-    "@angular/compiler-cli": "^21.0.3",
+    "@angular/build": "^21.1.3",
+    "@angular/cli": "^21.1.3",
+    "@angular/compiler-cli": "^21.1.3",
     "@types/jasmine": "~3.8.0",
-    "@types/node": "^12.11.1",
+    "@types/node": "^25.2.2",
     "jasmine-core": "~3.8.0",
     "karma": "~6.4.4",
     "karma-chrome-launcher": "~3.1.0",
@@ -45,5 +45,42 @@
     "karma-jasmine-html-reporter": "~1.7.0",
     "tailwindcss": "^3.4.17",
     "typescript": "~5.9.3"
+  },
+  "overrides": {
+    "glob": "^10.5.0",
+    "qs": "^6.14.1",
+    "tar": "^7.5.7",
+    "@isaacs/brace-expansion": "^5.0.1",
+    "@babel/helpers": "^7.26.10",
+    "lodash": "^4.17.23",
+    "brace-expansion": "^2.0.2",
+    "tmp": "^0.2.4",
+    "ws": "^8.18.0",
+    "webpack": "^5.105.0",
+    "webpack-dev-server": "^5.2.1",
+    "http-proxy-middleware": "^3.0.5",
+    "js-yaml": "^4.1.1",
+    "loader-utils": "^3.3.1",
+    "log4js": "^6.9.1",
+    "node-forge": "^1.3.2",
+    "on-headers": "^1.1.0",
+    "semver": "^7.7.1",
+    "socket.io-parser": "^4.2.4",
+    "engine.io": "^6.6.4",
+    "body-parser": "^2.2.0",
+    "express": "^5.1.0"
+  },
+  "pnpm": {
+    "overrides": {
+      "glob@>=10.2.0 <10.5.0": ">=10.5.0",
+      "qs@<6.14.1": ">=6.14.1",
+      "tar@<=7.5.6": ">=7.5.7",
+      "@isaacs/brace-expansion@<=5.0.0": ">=5.0.1",
+      "@babel/helpers@<7.26.10": ">=7.26.10",
+      "lodash@>=4.0.0 <=4.17.22": ">=4.17.23",
+      "brace-expansion@>=1.0.0 <=1.1.11": ">=1.1.12",
+      "brace-expansion@>=2.0.0 <=2.0.1": ">=2.0.2",
+      "tmp@<=0.2.3": ">=0.2.4"
+    }
   }
 }
