@@ -1,10 +1,10 @@
 import { Component, ChangeDetectionStrategy, computed, inject, signal } from '@angular/core';
-import {RouterOutlet, RouterLink, RouterLinkActive, Router, NavigationEnd} from '@angular/router';
-import {I18nService} from "../services/i18n/i18n.service";
-import {NgxSonnerToaster} from "ngx-sonner";
-import {filter} from 'rxjs/operators';
-import {LocalerouteService} from '../services/route/localeroute.service';
-import {ChatbotComponent} from './include/chatbot/chatbot.component';
+import { RouterOutlet, RouterLink, RouterLinkActive, Router, NavigationEnd } from '@angular/router';
+import { I18nService } from "../services/i18n/i18n.service";
+import { NgxSonnerToaster } from "ngx-sonner";
+import { filter } from 'rxjs/operators';
+import { LocalerouteService } from '../services/route/localeroute.service';
+import { ChatbotComponent } from './include/chatbot/chatbot.component';
 
 @Component({
   selector: 'app-root',
@@ -22,43 +22,37 @@ export class AppComponent {
   public isAdminRoute = signal<boolean>(false);
 
   constructor() {
-    // Detect language from initial URL
     const initialUrl = this.router.url;
     const langMatch = initialUrl.match(/^\/(fr|en)(\/|$)/);
     if (langMatch) {
       const lang = langMatch[1] as 'fr' | 'en';
       this.i18n.setLanguage(lang);
     }
-    // If no language prefix, the guard will handle it and use default language
 
     this.router.events.pipe(
       filter(event => event instanceof NavigationEnd)
     ).subscribe((event: any) => {
       this.isAdminRoute.set(event.url.startsWith('/admin'));
-      // Update language on navigation if URL has language prefix
       const urlMatch = event.url.match(/^\/(fr|en)(\/|$)/);
       if (urlMatch) {
         const lang = urlMatch[1] as 'fr' | 'en';
         this.i18n.setLanguage(lang);
       }
-      // Scroll to top on route change
       window.scrollTo({ top: 0, behavior: 'smooth' });
     });
-    // Check initial route
     this.isAdminRoute.set(this.router.url.startsWith('/admin'));
   }
 
   public languageImg = computed(() =>
     this.i18n.isEnglish()
-      ? 'assets/img/english.svg' // UK flag
-      : 'assets/img/france.svg' // France flag
+      ? 'assets/img/english.svg'
+      : 'assets/img/france.svg'
   );
 
   public currentLang = computed(() => this.i18n.isEnglish() ? 'en' : 'fr');
 
   private getCurrentPath(): string {
     const url = this.router.url;
-    // Remove language prefix if present
     const match = url.match(/^\/(fr|en)(\/.*)?$/);
     if (match) {
       return match[2] || '';
@@ -71,13 +65,11 @@ export class AppComponent {
     const langMatch = currentUrl.match(/^\/(fr|en)(\/.*)?$/);
 
     if (langMatch) {
-      // Current URL has language prefix - switch it
       const currentPath = langMatch[2] || '';
       const newLang = this.i18n.isEnglish() ? 'fr' : 'en';
       const newPath = `/${newLang}${currentPath}`;
       this.router.navigate([newPath]);
     } else {
-      // Current URL has no language prefix - just toggle language
       this.i18n.toggleLanguage();
     }
   }
