@@ -1,6 +1,6 @@
-import {NgModule} from '@angular/core';
-import {Title} from '@angular/platform-browser';
-import {provideHttpClient, withInterceptorsFromDi} from "@angular/common/http";
+import { NgModule } from '@angular/core';
+import { Title } from '@angular/platform-browser';
+import { provideHttpClient, withInterceptorsFromDi } from "@angular/common/http";
 
 @NgModule({
   providers: [
