@@ -1,5 +1,5 @@
-import {Component, OnInit} from '@angular/core';
-import {TempladminComponent} from "../../include/admin/templadmin/templadmin.component";
+import { Component, OnInit } from '@angular/core';
+import { TempladminComponent } from "../../include/admin/templadmin/templadmin.component";
 
 @Component({
   selector: 'app-admin',
