@@ -1,13 +1,13 @@
-import {NgModule} from '@angular/core';
-import {CommonModule} from '@angular/common';
-import {RouterModule, Routes} from "@angular/router";
-import {AdminComponent} from "./admin.component";
-import {AuthComponent} from "./auth/auth.component";
-import {WhitelistComponent} from "./whitelist/whitelist.component";
-import {AuthGuard} from "../../../guards/auth.guard";
-import {ProxiesComponent} from "./proxies/proxies.component";
-import {DetailserverComponent} from "./proxies/detailserver/detailserver.component";
-import {EmailComponent} from "./email/email.component";
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { RouterModule, Routes } from "@angular/router";
+import { AdminComponent } from "./admin.component";
+import { AuthComponent } from "./auth/auth.component";
+import { WhitelistComponent } from "./whitelist/whitelist.component";
+import { AuthGuard } from "../../../guards/auth.guard";
+import { ProxiesComponent } from "./proxies/proxies.component";
+import { DetailserverComponent } from "./proxies/detailserver/detailserver.component";
+import { EmailComponent } from "./email/email.component";
 
 const routes: Routes = [
   {
