@@ -10,7 +10,7 @@ describe('ErrorComponent', () => {
     await TestBed.configureTestingModule({
       imports: [ErrorComponent]
     })
-    .compileComponents();
+      .compileComponents();
 
     fixture = TestBed.createComponent(ErrorComponent);
     component = fixture.componentInstance;
