@@ -1,8 +1,8 @@
-import {NgModule} from '@angular/core';
-import {CommonModule} from '@angular/common';
-import {RouterModule, Routes} from "@angular/router";
-import {LegalComponent} from "./legal.component";
-import {CgvComponent} from "./cgv/cgv.component";
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { RouterModule, Routes } from "@angular/router";
+import { LegalComponent } from "./legal.component";
+import { CgvComponent } from "./cgv/cgv.component";
 
 const routes: Routes = [
   {
