@@ -1,6 +1,6 @@
-import {Routes} from "@angular/router";
-import {LegalComponent} from "./legal.component";
-import {CgvComponent} from "./cgv/cgv.component";
+import { Routes } from "@angular/router";
+import { LegalComponent } from "./legal.component";
+import { CgvComponent } from "./cgv/cgv.component";
 
 export const legalRoutes: Routes = [
   {
