@@ -1,6 +1,6 @@
-import {ComponentFixture, TestBed} from '@angular/core/testing';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import {TempladminComponent} from './templadmin.component';
+import { TempladminComponent } from './templadmin.component';
 
 describe('TempladminComponent', () => {
   let component: TempladminComponent;
