@@ -1,6 +1,6 @@
-import {ComponentFixture, TestBed} from '@angular/core/testing';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import {CompetencesComponent} from './competences.component';
+import { CompetencesComponent } from './competences.component';
 
 describe('CompetencesComponent', () => {
   let component: CompetencesComponent;
