@@ -1,6 +1,6 @@
-import {TestBed} from '@angular/core/testing';
+import { TestBed } from '@angular/core/testing';
 
-import {AdminService} from './admin.service';
+import { AdminService } from './admin.service';
 
 describe('AdminService', () => {
   let service: AdminService;
