@@ -1,12 +1,12 @@
-import {enableProdMode, provideZoneChangeDetection} from '@angular/core';
-import {bootstrapApplication} from '@angular/platform-browser';
-import {provideRouter, withComponentInputBinding} from '@angular/router';
-import {provideHttpClient, withInterceptorsFromDi} from '@angular/common/http';
-import {Title} from '@angular/platform-browser';
+import { enableProdMode, provideZoneChangeDetection } from '@angular/core';
+import { bootstrapApplication } from '@angular/platform-browser';
+import { provideRouter, withComponentInputBinding } from '@angular/router';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
+import { Title } from '@angular/platform-browser';
 
-import {AppComponent} from './app/components/app.component';
-import {routes} from './app/app-routing.module';
-import {environment} from './environments/environment';
+import { AppComponent } from './app/components/app.component';
+import { routes } from './app/app-routing.module';
+import { environment } from './environments/environment';
 
 if (environment.production) {
   enableProdMode();
@@ -14,7 +14,7 @@ if (environment.production) {
 
 bootstrapApplication(AppComponent, {
   providers: [
-    provideZoneChangeDetection(),provideRouter(routes, withComponentInputBinding()),
+    provideZoneChangeDetection(), provideRouter(routes, withComponentInputBinding()),
     provideHttpClient(withInterceptorsFromDi()),
     Title
   ]
